+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
  - Adopted [black](https://github.com/ambv/black) code style.
  - Removed auto-adding indexes, as it was unpredictable. You must add the `GinIndex` manually like described in the documentation on performance.
+ - Support dict for required_languages argument to TranslationField, to allow more fine-grained mapping of field names to required languages.
 
 
 ## 0.2.2 (2018-03-13)